pyproject.toml
[tool.poetry]
name = "balancy"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
web3 = "^5.17.0"
python-dotenv = "^0.15.0"
etcd3 = "^0.12.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
flake8 = "^3.8.4"
isort = "^5.6.4"
pre-commit = "^2.7.1"
pytest = "^6.1.1"
mypy = "^0.790"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
